If the root value is a collection, render the firstSummaryLine.
firstSummaryLine is the display that shows "Array (10) along with
the collapse/expand icon.

We're rendering this only inside a collection (so that the collection
key and the summary text can come in the same line).

To get it rendered for the root value, we need to handle it as a special
case, which this commit does. It feels hacky, but I think this duplication
is necessary to get this behaviour.