Rename “UniqueId” type as “NodePath”
The reason behind this change is that the “UniqueId” actually holds the absolute path of a node in a nested JSON tree. Absolute path to any node in a JSON tree is always unique, so “UniqueId” seems less informative. Instead, saying it as “NodePath” would be more appropriate. 

With this change, a “Node” now has “nodePath” property of “NodePath” type instead of “uniqueId” property.