Delimit a NodePath with dot character
We store absolute path of a node in a JSON (sub-)tree as the NodePath of a Node. The absolute path is formed by combining the keys (or indices) of each node at each level delimited by some delimiter. There’s no standard around, but many people follow to use a dot (“.”) character as the delimiter. So the resultant node path may look as “root.users.0.first_name” for example.

There might be a case where the key of a node in a JSON tree may contain hyphens, for example “first-name“; in such a case, if we use hyphen as delimiter the node path may not look as what we want as follows — “root-users-0-first-name”.