Ability to add parameters (#5)
* Add “More” actions dropdown next to URL input

Added “More” and “Add Parameter” options to the dropdown.
Added and fixed CSS style to properly fit the dropdown on the page.

* Handle change event “More” actions dropdown

Due to some browser’s non-compatibility for “onInput” event handler, it is advised to use custom “on” event for listening changes on a select field.

Registered “on” change event handler on the “More” actions dropdown to emit “MoreActionsDropdownChange” message with the selected option’s value.

Added “MoreActionsDropdownChange” value to “Msg” union type Msg module.

Revised the “update” function to handle the “Msg. MoreActionsDropdownChange” branch as well.
The “selectedOption” argument received by “MoreActionsDropdownChange” message is again could be one of many values. For now, provisioned it to return the model as is whenever user selects any dropdown option.

* Make the first option (“More”) selected after choosing any other option in the “More” actions dropdown

* Define “RequestParameter” and “RequestParameters” types

Added “RequestParameter” type to hold name and value of a request parameter.

Added “RequestParameters” type to hold an array of “RequestParameter” elements. The reason to use Array type is because we will need to update and delete the request parameter at a specific index and add a new one always at the tail. List and Set types seemed not appropriate for this problem, whereas, an Array type seemed to have nice support for it.

Added “requestParameters” property to the “Model” type.

Added an empty array for the “requestParameters” property while initializing the model in “init” constant.

* Introduce “AddRequestParameter” Msg

Added logic to handle the case when “Add Parameter” option is selected in “More” actions dropdown.
Provisioned it to emit “AddRequestParameter” message in such a case.

Added “AddRequestParameter” value to the Msg type in Msg module.

Added logic to handle when “AddRequestParameter” message is received in “update” function.
Provisioned it to push an empty “RequestParameter” element having blank name and value in the “requestParameters” property of the model.

* Render “requestParameters” stored in model

Added some view functions to help render the HTML markup for the request parameters stored in the model.
The “requestParameterItemView” function returns input fields for the name and value properties of a request parameter wrapped in a “LI” tag. 
The “requestParameterCollectionView” function returns an “UL” tag containing the list of input fields for the request parameters stored in the model.

Added a “requestParametersView” local variable in the main view function to render heading and the list of request parameters using above functions.

Added necessary CSS style to make it appear properly on the page.

* Add a request parameter on clicking “Add Parameter” link

Provisioned the “Add Parameter” link shown next to the “Request Parameters” heading to emit “AddRequestParameter” message on clicking on it.

* Update request parameters in model on input field change

Added “onInput” event handles on both “name” and “value” input fields for each rendered request parameter to emit “ChangeRequestParameterName” and “ChangeRequestParameterValue” messages with index as an argument respectively.

Added “ChangeRequestParameterName” and “ChangeRequestParameterValue” values to Msg type in Msg module.

Defined “RequestParameterPropertyName” union type to describe what are all properties are there in an element of type “RequestParameter”.

Added logic to update the specific (i.e. at a specified index) request parameter item’s property value with the new value in “updateRequestParameterProperty” function. It returns the updated model.

Added logic to handle messages viz. “ChangeRequestParameterName” and “ChangeRequestParameterValue” in main “update” function.

* Ability to delete a request parameter

As there’s core no support for removing an item at a given index in an Array, added “elm-community/array-extra” community package to the “elm-package.json” manifest.

Added a hyper link to delete every request parameter. Provisioned it to emit “DeleteRequestParameter” message with the index of the request parameter being deleted.

Added “DeleteRequestParameter” value to the Msg type in Msg module.

Added necessary CSS styling to display every request parameter delete hyperlink properly.

Provisioned the main “update” function to handle the received “DeleteRequestParameter” message and remove the requested request parameter at the specified index in the model.

* Attach request parameters as query parameters to the URL

Added a function “urlWithRequestParameters” which rewrites the URL by attaching the request parameters as query parameters.
While making a HTTP request, provisioned it to use such URL.

* Encode query parameter name and value

* Use Array.map instead of List.map

* Get rid of “updateRequestParameterProperty” function

As a consequence, the “Msg.ChangeRequestParameterName” and “Msg.ChangeRequestParameterName” branches now has duplicate code but which is okay as it avoids the need of an extra type “RequestParameterPropertyName”.

* Change signature of “urlWithRequestParameters” function

As per the name of function “urlWithRequestParameters”, flipped the arguments it accept in the following order: “url” and “requestParameters” instead of “requestParameters” and “url”.

* Rename "urlWithRequestParameters" function as "urlWithEncodedParameters"