Fix: Firebase App named '[DEFAULT]' already exists (app/duplicate-app)
If Firebase config is edited and saved by the user, we try to re-initialize the Firebase app using “firebase.initializeApp” SDK API.
As on page init, the default app is already initialized, it fails to re-initialize with this error.
To fix it, provisioned it to delete the previously initialized apps if any using the JS SDK, and re-initialize again.